requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
tqdm==4.66.5
aiohttp==3.13.3
